Implement CuraLine AI-powered appointment system
- Fix 9 critical bugs preventing app startup (app.setup(), broken DB session,
  LLM exception handler, Unicode schema typo, queue name mismatch, Alembic
  target_metadata, pydantic v1 import, Integer(5), datetime.now() default)
- Add DoctorSlot and RescheduleRequest models; update Patient and Appointment
  models with password_hash, severity_score, slot_id FK, and relationships
- Implement JWT authentication (register/login) with passlib/bcrypt
- Add full CRUD endpoints for doctors, slots, patients, and appointments
- Implement AI booking pipeline: multi-turn intake → severity analysis (1-5) →
  priority-based slot selection → booking → critical rescheduling requests
- Add reschedule accept/decline endpoints for patients asked to switch slots
- Add Celery periodic task to auto-close expired doctor slots
- Fix docker-compose: POSTGRES_DB env var, remove external network flag
- Add CORS middleware and global exception handlers to main.py

+++ models/doctor.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Enum, Integer, String, Time
+from sqlalchemy import Column, Enum, Integer, Numeric, String, Time
 
 from database.db import Base
 from models.mixins import PersonalDataMixin, TimestampMixin
@@ -18,7 +18,7 @@ class Doctor(Base, PersonalDataMixin, TimestampMixin):
     specialization = Column(String(100), nullable=False)
     qualification = Column(String(100), nullable=False)
     availability_status = Column(Enum(*AVAILABILITY_STATUSES), default=AVAILABLE)
-    consultation_fee = Column(Integer(5), nullable=False)
+    consultation_fee = Column(Numeric(7, 2), nullable=False)
     reporting_time = Column(Time, nullable=False)
     leaving_time = Column(Time, nullable=False)
     rating = Column(Integer, default=5)
